# be/docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.MeResponse:
    properties:
      role_name:
        example: admin
        type: string
      user_id:
        example: 20000000-0000-0000-0000-000000000001
        type: string
      username:
        example: admin
        type: string
    type: object
...
  models.CreateMenuRequest:
    properties:
      icon:
        example: bar-chart
        type: string
      name:
        example: Laporan
        type: string
      parent_id:
        type: string
      path:
        example: /reports
        type: string
      permission_id:
        type: string
      sort_order:
        example: 5
        type: integer
    required:
    - name
    type: object
...
  models.LoginRequest:
    properties:
      password:
        example: password123
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  models.LoginResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      expires_in:
        example: 86400
        type: integer
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      token_type:
        example: Bearer
        type: string
      user:
        $ref: '#/definitions/models.UserResponse'
    type: object
  models.Menu:
    properties:
      children:
        items:
          $ref: '#/definitions/models.Menu'
        type: array
      created_at:
        type: string
      icon:
        example: dashboard
        type: string
      id:
        example: 30000000-0000-0000-0000-000000000001
        type: string
      is_active:
        example: true
        type: boolean
      name:
        example: Dashboard
        type: string
      parent:
        $ref: '#/definitions/models.Menu'
      parent_id:
        type: string
      path:
        example: /dashboard
        type: string
      permission:
        $ref: '#/definitions/models.Permission'
      permission_id:
        type: string
      sort_order:
        example: 1
        type: integer
      updated_at:
        type: string
    type: object
  models.Permission:
    properties:
      code:
        example: user.view
        type: string
      created_at:
        type: string
      description:
        example: Can view user list and detail
        type: string
      id:
        example: a0000000-0000-0000-0000-000000000001
        type: string
      module:
        example: user
        type: string
      name:
        example: View Users
        type: string
      updated_at:
        type: string
    type: object
  models.RefreshTokenRequest:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
    required:
    - refresh_token
    type: object
  models.Role:
    properties:
      created_at:
        type: string
      description:
        example: Administrator with management access
        type: string
      id:
        example: 10000000-0000-0000-0000-000000000002
        type: string
      is_active:
        example: true
        type: boolean
      name:
        example: admin
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      updated_at:
        type: string
    type: object
  models.UpdateMenuRequest:
    properties:
      icon:
        example: bar-chart
        type: string
      is_active:
        example: true
        type: boolean
      name:
        example: Laporan Updated
        type: string
      parent_id:
        type: string
      path:
        example: /reports
        type: string
      permission_id:
        type: string
      sort_order:
        example: 5
        type: integer
    type: object
...
  models.UserResponse:
    properties:
      email:
        example: admin@example.com
        type: string
      full_name:
        example: Administrator
        type: string
      id:
        example: 20000000-0000-0000-0000-000000000002
        type: string
      is_active:
        example: true
        type: boolean
      role:
        $ref: '#/definitions/models.Role'
      username:
        example: admin
        type: string
    type: object
  utils.APIResponse:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/utils.Meta'
      success:
        type: boolean
    type: object
  utils.Meta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API dengan autentikasi lengkap, Dynamic RBAC, dan sistem menu
    hierarki.
  title: 247 Golang API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Autentikasi user dengan username dan password, mengembalikan access
        token dan refresh token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.LoginResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Login user
      tags:
      - Auth
  /auth/logout:
    post:
      description: Menghapus refresh token dan cache permission dari Redis
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - Auth
  /auth/me:
    get:
      description: Mendapatkan informasi user yang sedang login dari JWT token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handlers.MeResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Info user saat ini
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Mendapatkan access token baru menggunakan refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.LoginResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Refresh access token
      tags:
      - Auth
  /menus:
    get:
      description: Mendapatkan daftar semua menu tanpa hierarki (flat list). Memerlukan
        permission menu.view
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Menu'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Daftar semua menu (flat)
      tags:
      - Menus
    post:
      consumes:
      - application/json
      description: Membuat menu baru (bisa sebagai root atau sub-menu). Memerlukan
        permission menu.create
      parameters:
      - description: Data menu baru
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateMenuRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Menu'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Buat menu baru
      tags:
      - Menus
  /menus/{id}:
    delete:
      description: Menghapus menu berdasarkan ID. Memerlukan permission menu.delete
      parameters:
      - description: Menu ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Hapus menu
      tags:
      - Menus
    get:
      description: Mendapatkan detail menu beserta children-nya. Memerlukan permission
        menu.view
      parameters:
      - description: Menu ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Menu'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Detail menu
      tags:
      - Menus
    put:
      consumes:
      - application/json
      description: Mengupdate menu berdasarkan ID. Memerlukan permission menu.update
      parameters:
      - description: Menu ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Data update menu
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateMenuRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Menu'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update menu
      tags:
      - Menus
  /menus/my:
    get:
      description: Mendapatkan menu tree yang difilter berdasarkan permission user
        yang sedang login. Superadmin mendapatkan semua menu.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Menu'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Menu saya (Dynamic RBAC)
      tags:
      - Menus
  /menus/tree:
    get:
      description: Mendapatkan semua menu dalam format hierarki (tree). Memerlukan
        permission menu.view
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Menu'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Full menu tree
      tags:
      - Menus
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer {access_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"